Fix config watcher
